requirements.txt
# Core Dependencies
Django>=4.2.0
Pillow
python-dotenv
psycopg2-binary
redis==5.0.1
gunicorn==21.2.0

# Django Extensions
django-crispy-forms
crispy-bootstrap5
django-filter
djangorestframework>=3.14.0
django-storages==1.14.2
django-cors-headers
django-cleanup==8.0.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-redis==5.4.0
django-simple-history==3.4.0
django-widget-tweaks==1.5.0
django-allauth==0.61.1
django-ckeditor==6.7.0
django-compressor==4.4
django-environ==0.11.2
django-ratelimit
django-ses==3.5.0
django-taggit==5.0.1
django-tinymce==3.7.1

# WebSocket Support
channels
channels-redis
daphne==4.0.0

# Static Files
whitenoise

# Additional Dependencies
# celery>=5.3.0
drf-yasg>=1.21.0 